Netowrks Module

The Net module provides functionality for creating and manipulating power system network models in Julia. It includes features for defining buses, branches, transformers, prosumers, and shunts, as well as methods for running power flow analysis.

# Import the Net module
  using Sparlectra
  using BenchmarkTools
  # Create a network
  # Slack bus B1 at 220 kV with 1.0 pu voltage and 0.0 pu angle
  # PQ bus B2 at 220 kV with 1.0 pu voltage and 0.0 pu angle
  # PQ bus B3 at 22 kV with 1.0 pu voltage and 0.0 pu angle
  # Shunt at bus B3 with 0.0 kW and 150.0 kVar
  # AC line from bus B1 to B2 at 220 kV with 100 km length, 0.0653 ohm/km resistance, 0.398 ohm/km reactance, 9.08 nF/km capacitance, and 0.0 power factor
  # 2W transformer from bus B2 to B3 with 1000 MVA rating, 13.0% voltage ratio, 0.28% resistance, 20.0 kW power factor, and 0.06% no-load current
  # Energy consumer at bus B3 with 285.0 kW and 200.0 kVar
  # Synchronous machine at bus B1 with 1.02 pu voltage and 0.0 pu angle
  #
  # G->1* ----L---- 2 --T-- 3->L
  #                 x
  #   
  net = Net(name = "testnet", baseMVA = 100.0)
  addBus!(net = net, busName = "B1", busType = "Slack", vn_kV = 220.0, vm_pu = 1.0, va_deg = 0.0)
  addBus!(net = net, busName = "B2", busType = "PQ", vn_kV = 220.0, vm_pu = 1.0, va_deg = 0.0)
  addBus!(net = net, busName = "B3", busType = "PQ", vn_kV = 22.0, vm_pu = 1.0, va_deg = 0.0)
  addShunt!(net = net, busName = "B3", pShunt = 0.0, qShunt = 150.0)
  addACLine!(net = net, fromBus = "B1", toBus = "B2", length = 100.0, r = 0.0653, x = 0.398, c_nf_per_km = 9.08, tanδ = 0.0)
  add2WTrafo!(net = net, fromBus = "B2", toBus = "B3", sn_mva = 1000.0, vk_percent = 13.0, vkr_percent = 0.28, pfe_kw = 20.0, i0_percent = 0.06)
  addProsumer!(net = net, busName = "B3", type = "ENERGYCONSUMER", p = 285.0, q = 200.0)
  addProsumer!(net = net, busName = "B1", type = "SYNCHRONOUSMACHINE", referencePri = "B1", vm_pu = 1.02, va_deg = 0.0)

  # Run power flow
  tol = 1e-6
  maxIte = 10  
  etime = @elapsed begin
    ite, erg = runpf!(net, maxIte, tol, 0)
  end
  if erg != 0
    @warn "Power flow did not converge"        
  else
    calcNetLosses!(net)
    printACPFlowResults(net, etime, ite, tol)
  end
  

Sparlectra.Net — Type

Net: Represents an electrical network.

Fields:

name::String: Name of the network.

baseMVA::Float64: Base MVA of the network.

slackVec::Vector{Int}: Vector containing indices of slack buses.

vmin_pu::Float64: Minimum voltage limit in per unit.

vmax_pu::Float64: Maximum voltage limit in per unit.

nodeVec::Vector{Node}: Vector containing nodes in the network.

linesAC::Vector{ACLineSegment}: Vector containing AC line segments.

trafos::Vector{PowerTransformer}: Vector containing power transformers.

branchVec::Vector{Branch}: Vector containing branches in the network.

prosumpsVec::Vector{ProSumer}: Vector containing prosumers in the network.

shuntVec::Vector{Shunt}: Vector containing shunts in the network.

busDict::Dict{String,Int}: Dictionary mapping bus names to indices.

busOrigIdxDict::Dict{Int,Int}: Dictionary mapping current bus indices to original indices.

`branchDict::Dict{Tuple{Int, Int},Int} : Dictionary mapping branch tuples to indices.

totalLosses::Vector{Tuple{Float64,Float64}}: Vector containing tuples of total power losses.

_locked::Bool: Boolean indicating if the network is locked.

Constructors:

Net(name::String, baseMVA::Float64, vmin_pu::Float64 = 0.9, vmax_pu::Float64 = 1.1): Creates a new Net object with the given name, base MVA, and optional voltage limits.

Functions:

addBus!(; net::Net, ...): Adds a bus to the network.

addShunt!(; net::Net, ...): Adds a shunt to the network.

addBranch!(; net::Net, ...): Adds a branch to the network.

addACLine!(; net::Net, ...): Adds an AC line segment to the network.

addPIModellTrafo!(; net::Net, ...): Adds a transformer with PI model to the network.

add2WTrafo!(; net::Net, ...): Adds a two-winding transformer to the network.

addProsumer!(; net::Net, ...): Adds a prosumer to the network.

validate(; net::Net): Validates the network.

get_bus_vn_kV(; net::Net, busName::String): Gets the voltage level of a bus by name.

get_vn_kV(; net::Net, busIdx::Int): Gets the voltage level of a bus by index.

getBusType(; net::Net, busName::String): Gets the type of a bus by name.

lockNet!(; net::Net, locked::Bool): Locks or unlocks the network.

setBranchStatus!(; net::Net, branchIdx::Int, status::Int): Sets the status of a branch.

setTotalLosses!(; net::Net, pLosses::Float64, qLosses::Float64): Sets the total losses of the network.

getTotalLosses(; net::Net): Gets the total losses of the network.

Sparlectra.add2WTrafo! — Method

Add a two-winding transformer to the network.

Arguments

net::Net: The network to which the transformer will be added.

fromBus::String: The name of the bus where the transformer originates.

toBus::String: The name of the bus where the transformer terminates.

sn_mva::Float64: Rated power of the transformer.

vk_percent::Float64: Voltage regulation percent of the transformer.

vkr_percent::Float64: Voltage regulation percent of the transformer.

pfe_kw::Float64: Iron loss of the transformer.

i0_percent::Float64: No-load current percent of the transformer.

status::Int: The status of the transformer. Default is 1.

Sparlectra.addACLine! — Method

addACLine!: Adds an AC line segment to the network.

Parameters:

net::Net: Network object.

fromBus::String: Name of the "from" bus.

toBus::String: Name of the "to" bus.

length::Float64: Length of the line segment.

r::Float64: Resistance of the line segment.

x::Float64: Reactance of the line segment.

b::Union{Nothing,Float64} = nothing: Susceptance of the line segment (default is nothing).

c_nf_per_km::Union{Nothing,Float64} = nothing: Capacitance of the line segment in nF/km (default is nothing).

tanδ::Union{Nothing,Float64} = nothing: Tangent of the loss angle (default is nothing).

ratedS::Union{Nothing, Float64}= nothing: Rated power of the line segment in MVA (default is nothing).

status::Int = 1: Status of the line segment (default is 1).

Sparlectra.addBranch! — Method

addBranch!: Adds a branch to the network.

Parameters:

net::Net: Network object.

from::Int: Index of the "from" bus.

to::Int: Index of the "to" bus.

branch::AbstractBranch: Branch object to add.

status::Int = 1: Status of the branch (default is 1).

ratio::Union{Nothing,Float64} = nothing: Ratio of the branch (default is nothing).

side::Union{Nothing,Int} = nothing: Side of the branch (default is nothing).

vn_kV::Union{Nothing,Float64} = nothing: Nominal voltage of the branch in kV (default is nothing).

Sparlectra.addBus! — Method

addBus!: Adds a bus to the network.

Parameters:

net::Net: Network object.

busName::String: Name of the bus.

busType::String: Type of the bus (e.g., "Slack", "PQ", "PV").

vn_kV::Float64: Nominal voltage of the bus in kV.

vm_pu::Float64 = 1.0: Voltage magnitude of the bus in per unit (default is 1.0).

va_deg::Float64 = 0.0: Voltage angle of the bus in degrees (default is 0.0).

vmin_pu::Union{Nothing,Float64} = nothing: Minimum voltage limit in per unit (default is network's vmin_pu).

vmax_pu::Union{Nothing,Float64} = nothing: Maximum voltage limit in per unit (default is network's vmax_pu).

isAux::Bool = false: Boolean indicating if the bus is auxiliary (default is false).

oBusIdx::Union{Nothing,Int} = nothing: Original bus index (default is nothing).

zone::Union{Nothing,Int} = nothing: Zone index (default is nothing).

area::Union{Nothing,Int} = nothing: Area index (default is nothing).

ratedS::Union{Nothing,Float64} = nothing: Rated power of the bus in MVA (default is nothing).

Sparlectra.addPIModellTrafo! — Method

Add a transformer with PI model to the network.

Arguments

net::Net: The network to which the transformer will be added.

fromBus::String: The name of the bus where the transformer originates.

toBus::String: The name of the bus where the transformer terminates.

r_pu::Float64: The per-unit resistance of the transformer.

x_pu::Float64: The per-unit reactance of the transformer.

b_pu::Float64: The per-unit susceptance of the transformer.

status::Int: The status of the transformer.

ratedU::Union{Nothing, Float64}: Rated voltage of the transformer. Default is nothing.

ratedS::Union{Nothing, Float64}: Rated apparent power of the transformer. Default is nothing.

ratio::Union{Nothing, Float64}: Ratio of the transformer. Default is nothing.

shift_deg::Union{Nothing, Float64}: Phase shift angle of the transformer. Default is nothing.

isAux::Bool: Whether the transformer is an auxiliary transformer. Default is false.

Sparlectra.addProsumer! — Method

Add a prosumer (combination of a producer and consumer) to the network.

Arguments

net::Net: The network to which the prosumer will be added.

busName::String: The name of the bus where the prosumer is connected.

type::String: The type of the prosumer.

p::Union{Nothing, Float64}: Active power produced or consumed. Default is nothing.

q::Union{Nothing, Float64}: Reactive power produced or consumed. Default is nothing.

pMin::Union{Nothing, Float64}: Minimum active power. Default is nothing.

pMax::Union{Nothing, Float64}: Maximum active power. Default is nothing.

qMin::Union{Nothing, Float64}: Minimum reactive power. Default is nothing.

qMax::Union{Nothing, Float64}: Maximum reactive power. Default is nothing.

referencePri::Union{Nothing, String}: Reference bus for the prosumer. Default is nothing.

vm_pu::Union{Nothing, Float64}: Voltage magnitude setpoint. Default is nothing.

va_deg::Union{Nothing, Float64}: Voltage angle setpoint. Default is nothing.

Sparlectra.addShunt! — Method

addShunt!: Adds a shunt to the network.

Parameters:

net::Net: Network object.

busName::String: Name of the bus to which the shunt is added.

pShunt::Float64: Active power of the shunt in MW.

qShunt::Float64: Reactive power of the shunt in MVar.

in_service::Int = 1: Indicator for shunt's in-service status (default is 1).

Sparlectra.geNetBusIdx — Method

geNetBusIdx: Gets the index of a bus in the network.

Parameters:

net::Net: Network object.

busName::String: Name of the bus.

Returns:

Int: Index of the bus in the network.

Sparlectra.getBusType — Method

Get the type of a specific bus in the network.

Arguments

net::Net: The network from which to retrieve the bus type.

busName::String: The name of the bus.

Returns

The type of the specified bus.

Sparlectra.getNetOrigBusIdx — Method

getNetOrigBusIdx: Gets the original index of a bus in the network.

Parameters:

net::Net: Network object.

busName::String: Name of the bus.

Returns:

Int: Original index of the bus in the network.

Sparlectra.getTotalLosses — Method

Get the total losses in the network.

Arguments

net::Net: The network from which to retrieve the losses.

Returns

A tuple (pLosses::Float64, qLosses::Float64) containing the total active and reactive power losses in the network.

Sparlectra.get_bus_vn_kV — Method

Get the voltage magnitude of a specific bus in the network.

Arguments

net::Net: The network from which to retrieve the voltage magnitude.

busName::String: The name of the bus.

Returns

The voltage magnitude of the specified bus.

Sparlectra.get_vn_kV — Method

Get the voltage magnitude of a specific bus in the network.

Arguments

net::Net: The network from which to retrieve the voltage magnitude.

busIdx::Int: The index of the bus.

Returns

The voltage magnitude of the specified bus.

Sparlectra.hasBusInNet — Method

hasBusInNet: Checks if a bus exists in the network.

Parameters:

net::Net: Network object.

busName::String: Name of the bus to check.

Returns:

Bool: True if the bus exists in the network, otherwise false.

Sparlectra.lockNet! — Method

Lock or unlock the network.

Arguments

net::Net: The network to be locked or unlocked.

locked::Bool: Boolean indicating whether to lock the network.

Sparlectra.setBranchStatus! — Method

Set the status of a branch in the network.

Arguments

net::Net: The network object.

fromBus::String: The name of the bus where the branch originates.

toBus::String: The name of the bus where the branch terminates.

status::Int: The new status of the branch.

Examples

net = run_acpflow(max_ite= 7,tol = 1e-6, casefile='a_case.m') # run the power flow on the network and get the network object
setBranchStatus!(net, "Bus1", "Bus2", 1)  # Set the status of the branch from Bus1 to Bus2 to 1.
run_net_acpflow(net = net, max_ite= 7,tol = 1e-6) # rerun the power flow with the updated network

Sparlectra.setTotalLosses! — Method

Set the total losses in the network.

Arguments

net::Net: The network to which the losses will be added.

pLosses::Float64: Total active power losses.

qLosses::Float64: Total reactive power losses.

Sparlectra.validate — Method

Validate the network configuration.

Arguments

net::Net: The network to be validated.

Returns

A tuple (valid::Bool, message::String) where valid is a boolean indicating whether the network is valid, and message is a string containing an error message if the network is invalid.