Function Reference

Sparlectra.adjacentBranches — Function

adjacentBranches: Find adjacent branches for each node in the network.

Parameters:

Y::AbstractMatrix{ComplexF64}: Admittance matrix of the network.

log::Bool = false: Optional parameter indicating whether to print the adjacent branches (default is false).

Returns:

adjList::Vector{Vector{Int}}: Vector of vectors containing the indices of adjacent branches for each node.

Sparlectra.createYBUS — Function

createYBUS: Create the admittance matrix YBUS from the branch vector and the slack index.

Parameters:

branchVec::Vector{Branch}: Vector of branch objects representing the network branches.

shuntVec::Vector{Shunt}: Vector of shunt objects representing shunt elements in the network.

sparse::Bool = true: Optional parameter indicating whether to create a sparse YBUS matrix (default is true).

printYBUS::Bool = false: Optional parameter indicating whether to print the YBUS matrix (default is false).

Returns:

Y::SparseMatrixCSC{ComplexF64} or Array{ComplexF64}: The YBUS admittance matrix.

Sparlectra.runpf! — Function

Runs the power flow calculation using the Newton-Raphson method.

Arguments

net::Net: Network data structure.

maxIte::Int: Maximum number of iterations.

tolerance::Float64 = 1e-6: Tolerance for convergence (default: 1e-6).

verbose::Int = 0: Verbosity level (default: 0).

Returns

A tuple containing the number of iterations and the result of the calculation:

0: Convergence reached.

1: No convergence.

2: Unsolvable system of equations.

3: Error.